requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
matplotlib>=3.7.0 

# Data Processing
openpyxl>=3.1.2
xlrd>=2.0.1
pyarrow>=13.0.0
feather-format>=0.4.1

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.9

# Statistical Analysis
scipy>=1.11.0
scikit-learn>=1.3.0
statsmodels>=0.14.0

# API and HTTP
requests>=2.31.0
httpx>=0.25.0

# Environment and Configuration
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

statsmodels
